Django/requirements.txt
Django==5.0.7
django_auditlog==2.3.0
django-filter==24.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==26.0.0
flake8==7.1.0
gunicorn==21.2.0
model-bakery==1.18.2
requests==2.31.0
selenium==4.18.1
slippers==0.6.2
psycopg2-binary==2.9.9
yfinance==0.2.36
